Do not block on database inserts and reduce logspam
Increases peak inserts/sec by up to ~13x.

We also silence the warning about missed windows, because the logging takes up
a significant portion of CPU time. With the current design, missed windows can
never be prevented completely and should be considered a normal part of
operation (at least for the trace-service). Number of dropped records can still
be seen in the Grafana dashboard.