Raise Cassandra connection limits to avoid failures
With the previous throughput improvement, we are now hitting the default query limit
of the DataStax Cassandra driver, leading to All nodes unavailable errors for
some queries.

These settings may not be optimal but they resolve the failures for now. In particular,
more distributed Cassandra nodes would be preferable to many connections per node.