Raise kafka-streams message size limits
In tests with >10,000 spans/sec we hit kafka message size limits, crashing the
server after a few minutes with `Error encountered sending record to topic trace-service-KTABLE-SUPPRESS-STATE-STORE-0000000009-changelog for task 1_10 due to: org.apache.kafka.common.errors.RecordTooLargeException: The request included a message larger than the max message size the server will accept.`

The new size was chosen arbitrarily, since to my knowledge we do not even send
messages this big and therefore this issue shouldn't occur, but the new limit
seems to fix the issue in this particular instance.

Needs a corresponding change in the broker configuration (see MR).